models: move models package
